models/vae_training_code/requirements.txt
torch
torchvision
diffusers
transformers
accelerate
datasets
xarray
h5netcdf
hydra-core
wandb